extract RadarSDKFraud getFraudPayload
