+++ sdk/src/test/java/io/radar/sdk/RadarRevealRiskTest.kt
@@ -1,0 +1,16 @@
+package io.radar.sdk
+
+import android.os.Build
+import androidx.test.ext.junit.runners.AndroidJUnit4
+import org.junit.runner.RunWith
+import org.robolectric.annotation.Config
+import kotlin.intArrayOf
+
+
+@RunWith(AndroidJUnit4::class)
+@Config(sdk = [Build.VERSION_CODES.P])
+class RadarRevealRiskTest {
+
+
+
+}
